Add exceptions handling; Rename models to schemas
